+++ src/websocket.cpp
@@ -4,7 +4,7 @@
 
 // WebRTC
 #include <rtc_base/logging.h>
-#include <rtc_base/third_party/base64/base64.h>
+#include <rtc_base/base64.h>
 
 // Boost
 #include <boost/asio/bind_executor.hpp>
@@ -394,7 +394,7 @@ void Websocket::OnConnectProxy(boost::system::error_code ec) {
   proxy_req_.set(boost::beast::http::field::host, target);
   proxy_req_.set(
       boost::beast::http::field::proxy_authorization,
-      "Basic " + rtc::Base64::Encode(proxy_username_ + ":" + proxy_password_));
+      "Basic " + webrtc::Base64Encode(proxy_username_ + ":" + proxy_password_));
   boost::beast::http::async_write(
       *proxy_socket_, proxy_req_,
       std::bind(&Websocket::OnWriteProxy, this, std::placeholders::_1,